LayoutTests/fast/scrolling/mac/rubberband-overflow-with-non-scrollable-root.html

<!DOCTYPE html> <!-- webkit-test-runner [ AsyncOverflowScrollingEnabled=true ] -->
<html>
<head>
    <style>
        body {
            height: 2000px;
            overflow: hidden;
        }

        .scroller {
            height: 300px;
            width: 300px;
            border: 1px solid black;
            padding: 5px;
            overflow: scroll;
        }
        .content {
            width: 200%;
            height: 300%;
        }
        
    </style>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <script>
        var jsTestIsAsync = true;

        var scroller;
        var sawOverflowScrollWithNegativeOffset = false;

        async function resetScrollPositions()
        {
            scroller.scrollTop = 0;
            
            // Wait for scroll events to fire.
            await UIHelper.renderingUpdate();

            sawOverflowScrollWithNegativeOffset = false;
        }
        
        async function doMouseWheelScroll()
        {
            eventSender.mouseMoveTo(150, 200);

            // Pull down.
            const wheelEventSquence = {
                "events" : [
                    {
                        type : "wheel",
                        viewX : 150,
                        viewY : 200,
                        deltaY : 1,
                        phase : "began"
                    },
                    {
                        type : "wheel",
                        deltaY : 10,
                        phase : "changed"
                    },
                    {
                        type : "wheel",
                        viewX : 201, // defeat coalescing
                        deltaY : 10,
                        phase : "changed"
                    },
                    {
                        type : "wheel",
                        phase : "ended"
                    }
                ]
            };

            await UIHelper.sendEventStream(wheelEventSquence);
            await UIHelper.renderingUpdate();
        }
        
        async function scrollTest()
        {
            debug('');
            debug('Pulling down in the overflow should rubberband it.');
            await resetScrollPositions();

            await doMouseWheelScroll();

            shouldBecomeEqual('sawOverflowScrollWithNegativeOffset', 'true', finishJSTest);
        }

        window.addEventListener('load', () => {
            scroller = document.querySelector('.scroller');
            scroller.addEventListener('scroll', () => {
                if (scroller.scrollTop < 0)
                    sawOverflowScrollWithNegativeOffset = true;
            }, false);

            setTimeout(scrollTest, 0);
        }, false);
    </script>
</head>
<body>
    <div class="scroller">
        <div class="content">Contents</div>
    </div>
    <script src="../../../resources/js-test-post.js"></script>
</body>
</html>
